reorganize render process
