+++ packages/cli/src/render.tsx
@@ -10,12 +10,10 @@ import chalk from 'chalk';
 import fs from 'fs';
 import os from 'os';
 import path from 'path';
-import {Config, Internals} from 'remotion';
+import {Internals} from 'remotion';
+import {getCliOptions} from './get-cli-options';
 import {getCompositionId} from './get-composition-id';
 import {getConfigFileName} from './get-config-file-name';
-import {getOutputFilename} from './get-filename';
-import {getInputProps} from './get-input-props';
-import {getImageFormat} from './image-formats';
 import {loadConfigFile} from './load-config';
 import {Log} from './log';
 import {parseCommandLine, parsedCli} from './parse-command-line';
@@ -25,8 +23,7 @@ import {
 	makeRenderingProgress,
 	makeStitchingProgres,
 } from './progress-bar';
-import {getUserPassedFileExtension} from './user-passed-output-location';
-import {warnAboutFfmpegVersion} from './warn-about-ffmpeg-version';
+import {checkAndValidateFfmpegVersion} from './validate-ffmpeg-version';
 
 export const render = async () => {
 	const startTime = Date.now();
@@ -39,105 +36,26 @@ export const render = async () => {
 	if (appliedName) {
 		Log.Verbose(`Applied configuration from ${appliedName}.`);
 	}
-	const parallelism = Internals.getConcurrency();
-	const frameRange = Internals.getRange();
-	if (typeof frameRange === 'number') {
-		Log.Warn('Selected a single frame. Assuming you want to output an image.');
-		Log.Warn(
-			`If you want to render a video, pass a range:  '--frames=${frameRange}-${frameRange}'.`
-		);
-		Log.Warn("To dismiss this message, add the '--sequence' flag explicitly.");
-		Config.Output.setImageSequence(true);
-	}
-	const shouldOutputImageSequence = Internals.getShouldOutputImageSequence();
-	const userCodec = Internals.getOutputCodecOrUndefined();
+	const {
+		codec,
+		parallelism,
+		frameRange,
+		shouldOutputImageSequence,
+		absoluteOutputFile,
+		overwrite,
+		inputProps,
+		quality,
+		browser,
+		crf,
+		pixelFormat,
+		imageFormat,
+	} = await getCliOptions();
 
-	const codec = Internals.getFinalOutputCodec({
-		codec: userCodec,
-		fileExtension: getUserPassedFileExtension(),
-		emitWarning: true,
-	});
+	checkAndValidateFfmpegVersion();
 
-	const ffmpegVersion = await RenderInternals.getFfmpegVersion();
-	Log.Verbose(
-		'Your FFMPEG version:',
-		ffmpegVersion ? ffmpegVersion.join('.') : 'Built from source'
-	);
-	warnAboutFfmpegVersion(ffmpegVersion);
-	if (
-		codec === 'vp8' &&
-		!(await RenderInternals.ffmpegHasFeature('enable-libvpx'))
-	) {
-		Log.Error(
-			"The Vp8 codec has been selected, but your FFMPEG binary wasn't compiled with the --enable-lipvpx flag."
-		);
-		Log.Error(
-			'This does not work, please switch out your FFMPEG binary or choose a different codec.'
-		);
-	}
-	if (
-		codec === 'h265' &&
-		!(await RenderInternals.ffmpegHasFeature('enable-gpl'))
-	) {
-		Log.Error(
-			"The H265 codec has been selected, but your FFMPEG binary wasn't compiled with the --enable-gpl flag."
-		);
-		Log.Error(
-			'This does not work, please recompile your FFMPEG binary with --enable-gpl --enable-libx265 or choose a different codec.'
-		);
-	}
-	if (
-		codec === 'h265' &&
-		!(await RenderInternals.ffmpegHasFeature('enable-libx265'))
-	) {
-		Log.Error(
-			"The H265 codec has been selected, but your FFMPEG binary wasn't compiled with the --enable-libx265 flag."
-		);
-		Log.Error(
-			'This does not work, please recompile your FFMPEG binary with --enable-gpl --enable-libx265 or choose a different codec.'
-		);
-	}
-
-	const outputFile = getOutputFilename(codec, shouldOutputImageSequence);
-	const overwrite = Internals.getShouldOverwrite();
-	const inputProps = getInputProps();
-	const quality = Internals.getQuality();
-	const browser = Internals.getBrowser() ?? Internals.DEFAULT_BROWSER;
 	const browserInstance = RenderInternals.openBrowser(browser, {
 		shouldDumpIo: Internals.Logging.isEqualOrBelowLogLevel('verbose'),
 	});
-
-	const absoluteOutputFile = path.resolve(process.cwd(), outputFile);
-	if (fs.existsSync(absoluteOutputFile) && !overwrite) {
-		Log.Error(
-			`File at ${absoluteOutputFile} already exists. Use --overwrite to overwrite.`
-		);
-		process.exit(1);
-	}
-	if (!shouldOutputImageSequence) {
-		await RenderInternals.validateFfmpeg();
-	}
-	const crf = shouldOutputImageSequence ? null : Internals.getActualCrf(codec);
-	if (crf !== null) {
-		Internals.validateSelectedCrfAndCodecCombination(crf, codec);
-	}
-	const pixelFormat = Internals.getPixelFormat();
-	const imageFormat = getImageFormat(
-		shouldOutputImageSequence ? undefined : codec
-	);
-
-	Internals.validateSelectedPixelFormatAndCodecCombination(pixelFormat, codec);
-	Internals.validateSelectedPixelFormatAndImageFormatCombination(
-		pixelFormat,
-		imageFormat
-	);
-	try {
-		await RenderInternals.ensureLocalBrowser(browser);
-	} catch (err) {
-		Log.Error('Could not download a browser for rendering frames.');
-		Log.Error(err);
-		process.exit(1);
-	}
 	if (shouldOutputImageSequence) {
 		fs.mkdirSync(absoluteOutputFile, {
 			recursive: true,
@@ -179,7 +97,7 @@ export const render = async () => {
 	}
 	const openedBrowser = await browserInstance;
 	const comps = await getCompositions(bundled, {
-		browser: Internals.getBrowser() || Internals.DEFAULT_BROWSER,
+		browser,
 		inputProps,
 		browserInstance: openedBrowser,
 	});
